refactor: 게시물 업로드 컴포넌트 분리 및 구조 개선
