+++ src/pages/labs/LabView.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import BasePostView from '../../components/posts/BasePostView';
+import BasePostView from '../../components/posts/view/BasePostView';
 
 function LabView() {
     return <BasePostView collectionName="labs" />;
